Fix some OS compatibility issues
On Windows, Python scripts must be executed by the 'python'
interpreter but on Linux, they must be executed by the 'python3'
interpreter.

Now, to choose wich one to use, AcHelper search if /usr/bin/python3
exists.

If so, the 'python3' interpreter is used, otherwise it is the 'python'
interpreter.

Python files will always be executed by the right version of python
interpreter.